Refactored WGSL parser into Nano.h to keep as much contained within headers as possible. Refactored custom font loading to now require the user to either define their custom fonts before nano_start_app() is called in main() or Nano will use no custom fonts. This means ImGui will have the default font if we are using nano_cimgui.h, and Nano will have no fonts that it can render directly to the frame.
