+++ include/CMakeLists.txt
@@ -1,3 +1,2 @@
 # add_subdirectory(cglm)
 add_subdirectory(cimgui)
-add_subdirectory(wgsl-parser)
