--- include/wgsl-parser/CMakeLists.txt
@@ -1,24 +1,0 @@
-cmake_minimum_required(VERSION 3.12)
-project(wgsl-parser)
-
-# Create the wgsl_parser library
-add_library(wgsl-parser STATIC
-    wgsl-parser.c
-    wgsl-parser.h
-    webgpu.h
-)
-
-# Create the test executable
-# Make sure it creates a binary called test
-add_executable(test
-    test.c
-)
-
-
-# # Link the wgsl_parser library to the test executable
-target_link_libraries(test PRIVATE wgsl-parser)
-
-# Remove the files generated by Emscripten using clean
-set_property(DIRECTORY PROPERTY ADDITIONAL_MAKE_CLEAN_FILES
-    "test.js;test.wasm;"
-)
