--- include/wgsl-parser/README.md
@@ -1,3 +1,0 @@
-# wgsl-parser
-
-Correct binding definition parsing so that it can test for multiple regex patterns to cover many cases.
